Refactor timestamp parsing into reusable helper
